Fix the WinFormTranslator crash
DODGY HACK!!

I suspect we possibly don't actually need the original Translator.Instance anymore, but I'll leave it for now (it's not what I am focused on)
Anyway... you can't just cast to WinFormTranslator here because the incoming translator could (and indeed is!) of type Translator.
So instead, I've created a second singleton specifically for WinFormTranslator - and use THAT instead of the incoming translator.